Ready tests for GetDirectoryPathFromFilename

+++ src/Utilities/Directory/DeleteDirectoryWithAllContents.php
@@ -1,11 +1,12 @@
 <?php
+
 declare(strict_types=1);
 
 namespace Ifrost\Common\Utilities\Directory;
 
-use Ifrost\Common\HandleInterface;
+use Ifrost\Common\Executable;
 
-class DeleteDirectoryWithAllContents implements HandleInterface
+class DeleteDirectoryWithAllContents implements Executable
 {
     private string $path;
 
@@ -14,7 +15,7 @@ class DeleteDirectoryWithAllContents implements HandleInterface
         $this->path = $path;
     }
 
-    public function handle(): void
+    public function execute(): void
     {
         $this->delete($this->path);
     }
@@ -22,24 +23,28 @@ class DeleteDirectoryWithAllContents implements HandleInterface
     private function delete(string $dirPath): void
     {
         if (!is_dir($dirPath)) {
-            throw new \InvalidArgumentException(sprintf("%s is not directory.", $dirPath));
+            throw new \InvalidArgumentException(sprintf('%s is not directory.', $dirPath));
         }
 
-        if (substr($dirPath, strlen($dirPath) - 1, 1) != '/') {
+        if (substr($dirPath, strlen($dirPath) - 1, 1) !== '/') {
             $dirPath .= '/';
         }
 
         $files = glob($dirPath . '*', GLOB_MARK);
 
+        if ($files === false) {
+            throw new \RuntimeException(sprintf('Unable to find files inside "%s".', $dirPath));
+        }
+
         foreach ($files as $file) {
             if (is_dir($file)) {
                 $this->delete($file);
-            } else {
-                try {
-                    unlink($file);
-                } catch (\Exception $e) {
-                    throw new \RuntimeException(sprintf('Unable remove file "%s".', $file));
-                }
+
+                continue;
+            }
+
+            if (!unlink($file)) {
+                throw new \RuntimeException(sprintf('Unable remove directory "%s".', $dirPath));
             }
         }
